services/data-api-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
neo4j>=5.8.0
asyncpg>=0.27.0
httpx>=0.24.0
